Extract ParseDuration to Duration class
- Make Duration.isEqualTo null-safe
